+++ requirements.txt
@@ -3,6 +3,7 @@ aiohttp==3.12.13
 aiosignal==1.3.2
 aiosqlite==0.21.0
 alembic==1.16.2
+altair==5.5.0
 annotated-types==0.7.0
 anyio==4.9.0
 appnope==0.1.4
@@ -18,6 +19,7 @@ attrs==25.3.0
 babel==2.17.0
 beautifulsoup4==4.13.4
 bleach==6.2.0
+blinker==1.9.0
 cachetools==5.5.2
 certifi==2025.6.15
 cffi==1.17.1
@@ -49,6 +51,8 @@ fqdn==1.5.1
 frozendict==2.4.6
 frozenlist==1.7.0
 fsspec==2025.5.1
+gitdb==4.0.12
+GitPython==3.1.44
 google-api-core==2.25.1
 google-auth==2.40.3
 google-auth-oauthlib==1.2.2
@@ -137,7 +141,6 @@ orjson==3.10.18
 overrides==7.7.0
 packaging==24.2
 pandas==2.3.0
-pandas-gbq==0.29.1
 pandocfilters==1.5.1
 parso==0.8.4
 pathspec==0.12.1
@@ -168,6 +171,7 @@ pydantic-extra-types==2.10.5
 pydantic-settings==2.9.1
 pydantic_core==2.33.2
 pydata-google-auth==1.9.1
+pydeck==0.9.1
 Pygments==2.19.1
 pyparsing==3.2.3
 python-dateutil==2.9.0.post0
@@ -198,11 +202,13 @@ Send2Trash==1.8.3
 sentence-transformers==4.1.0
 shellingham==1.5.4
 six==1.17.0
+smmap==5.0.2
 sniffio==1.3.1
 soupsieve==2.7
 SQLAlchemy==2.0.41
 stack-data==0.6.3
 starlette==0.46.2
+streamlit==1.46.0
 sympy==1.14.0
 tenacity==9.1.2
 terminado==0.18.1
@@ -228,6 +234,7 @@ uri-template==1.3.0
 urllib3==2.5.0
 uv==0.7.13
 uvicorn==0.34.3
+watchdog==6.0.0
 wcwidth==0.2.13
 webcolors==24.11.1
 webencodings==0.5.1
